requirements.txt
django
djangorestframework
psycopg2-binary
django-bootstrap_datepicker_plus
django-bootstrap4
drf-yasg